Prima creazione Labirinto
