+++ src/DiaDia.java
@@ -93,16 +93,16 @@ public class DiaDia {
 			System.out.println("Dove vuoi andare ?");
 		Stanza prossimaStanza = null;
 		
-		prossimaStanza = this.partita.getStanzaCorrente().getStanzaAdiacente(direzione);
+		prossimaStanza = this.partita.getLabirinto().getStanzaCorrente().getStanzaAdiacente(direzione);
 		
 		if (prossimaStanza == null)
 			System.out.println("Direzione inesistente");
 		else {
-			this.partita.setStanzaCorrente(prossimaStanza);
+			this.partita.getLabirinto().setStanzaCorrente(prossimaStanza);
 			int cfu = this.partita.getCfu();
 			this.partita.setCfu(--cfu);
 		}
-		System.out.println(partita.getStanzaCorrente().getDescrizione());
+		System.out.println(this.partita.getLabirinto().getStanzaCorrente().getDescrizione());
 		
 	}
