+++ src/Partita.java
@@ -11,15 +11,14 @@
 
 public class Partita {
 
-	static final private int CFU_INIZIALI = 20;
-
-	private Stanza stanzaCorrente;
-	private Stanza stanzaVincente;
+	private static final int CFU_INIZIALI = 20;
+	
+	private Labirinto labirinto;
 	private boolean finita;
 	private int cfu;
 	
 	public Partita(){
-		creaStanze();
+		Labirinto.creaStanze();
 		this.finita = false;
 		this.cfu = CFU_INIZIALI;
 	}
@@ -27,60 +26,14 @@ public class Partita {
     /**
      * Crea tutte le stanze e le porte di collegamento
      */
-    private void creaStanze() {
-
-		/* crea gli attrezzi */
-    	Attrezzo lanterna = new Attrezzo("lanterna",3);
-		Attrezzo osso = new Attrezzo("osso",1);
-    	
-		/* crea stanze del labirinto */
-		Stanza atrio = new Stanza("Atrio");
-		Stanza aulaN11 = new Stanza("Aula N11");
-		Stanza aulaN10 = new Stanza("Aula N10");
-		Stanza laboratorio = new Stanza("Laboratorio Campus");
-		Stanza biblioteca = new Stanza("Biblioteca");
-		
-		/* collega le stanze */
-		atrio.impostaStanzaAdiacente("nord", biblioteca);
-		atrio.impostaStanzaAdiacente("est", aulaN11);
-		atrio.impostaStanzaAdiacente("sud", aulaN10);
-		atrio.impostaStanzaAdiacente("ovest", laboratorio);
-		aulaN11.impostaStanzaAdiacente("est", laboratorio);
-		aulaN11.impostaStanzaAdiacente("ovest", atrio);
-		aulaN10.impostaStanzaAdiacente("nord", atrio);
-		aulaN10.impostaStanzaAdiacente("est", aulaN11);
-		aulaN10.impostaStanzaAdiacente("ovest", laboratorio);
-		laboratorio.impostaStanzaAdiacente("est", atrio);
-		laboratorio.impostaStanzaAdiacente("ovest", aulaN11);
-		biblioteca.impostaStanzaAdiacente("sud", atrio);
-
-        /* pone gli attrezzi nelle stanze */
-		aulaN10.addAttrezzo(lanterna);
-		atrio.addAttrezzo(osso);
-
-		// il gioco comincia nell'atrio
-        stanzaCorrente = atrio;  
-		stanzaVincente = biblioteca;
-    }
-
-	public Stanza getStanzaVincente() {
-		return stanzaVincente;
-	}
-
-	public void setStanzaCorrente(Stanza stanzaCorrente) {
-		this.stanzaCorrente = stanzaCorrente;
-	}
-
-	public Stanza getStanzaCorrente() {
-		return this.stanzaCorrente;
-	}
+    
 	
 	/**
 	 * Restituisce vero se e solo se la partita e' stata vinta
 	 * @return vero se partita vinta
 	 */
 	public boolean vinta() {
-		return this.getStanzaCorrente()== this.getStanzaVincente();
+		return getLabirinto().getStanzaCorrente()== Labirinto.getStanzaFinale();
 	}
 
 	/**
@@ -90,7 +43,7 @@ public class Partita {
 	public boolean isFinita() {
 		return finita || vinta() || (cfu == 0);
 	}
-
+	
 	/**
 	 * Imposta la partita come finita
 	 *
@@ -98,7 +51,11 @@ public class Partita {
 	public void setFinita() {
 		this.finita = true;
 	}
-
+	
+	public Labirinto getLabirinto() {
+		return this.labirinto;
+	}
+	
 	public int getCfu() {
 		return this.cfu;
 	}
